04_langchain/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
langchain
langchain-community
openai
PyPDF2
python-dotenv
langchain-google-genai